# server/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "L5 Swagger OpenApi description",
        "version": "1.0.0"
    },
    "paths": {
        "/api/favorites": {
            "get": {
                "tags": [
                    "Favorites"
                ],
                "summary": "Lista os filmes favoritos",
                "operationId": "d0d405a490978063e03bab284ca93c18",
                "responses": {
                    "200": {
                        "description": "Lista de filmes favoritos retornada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id_tmdb": {
                                                "type": "integer"
                                            },
                                            "movie_title": {
                                                "type": "string"
                                            },
                                            "poster_path": {
                                                "type": "string"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Favorites"
                ],
                "summary": "Adiciona um filme aos favoritos",
                "operationId": "38ecd9fddd24c32aa0704ad440f5cf9d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_tmdb"
                                ],
                                "properties": {
                                    "id_tmdb": {
                                        "description": "ID do filme no TMDB",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Filme adicionado aos favoritos com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id_tmdb": {
                                            "type": "integer"
                                        },
                                        "movie_title": {
                                            "type": "string"
                                        },
                                        "overview": {
                                            "type": "string"
                                        },
                                        "poster_path": {
                                            "type": "string"
                                        },
                                        "realease_date": {
                                            "type": "string"
                                        },
                                        "rating": {
                                            "type": "number",
                                            "format": "float"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                }
            }
        },
        "/api/favorites/{id}": {
            "delete": {
                "tags": [
                    "Favorites"
                ],
                "summary": "Remove um filme dos favoritos",
                "operationId": "0cf69332240f76f1c7967d2fd6f8a3dc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do filme",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Filme removido dos favoritos com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Filme não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/movies/list-top-rateds": {
            "get": {
                "tags": [
                    "Movies"
                ],
                "summary": "Retorna filmes mais bem avaliados",
                "operationId": "02378cb8630cced8f002c749426681c6",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número da página para paginação",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de filmes mais bem avaliados retornada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "results": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "page": {
                                            "type": "integer"
                                        },
                                        "total_pages": {
                                            "type": "integer"
                                        },
                                        "total_results": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro ao buscar filmes"
                    }
                }
            }
        },
        "/api/movies/searchByName": {
            "get": {
                "tags": [
                    "Movies"
                ],
                "summary": "Busca filmes na API do TMDB com base no nome informado",
                "operationId": "ac8529986d872372aadab988dd933722",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Nome do filme para buscar",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resultados da busca retornados com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "results": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "page": {
                                            "type": "integer"
                                        },
                                        "total_pages": {
                                            "type": "integer"
                                        },
                                        "total_results": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro: Nome não informado"
                    }
                }
            }
        },
        "/api/movies/searchById/{id_tmdb}": {
            "get": {
                "tags": [
                    "Movies"
                ],
                "summary": "Busca detalhes de um filme pelo ID do TMDB",
                "operationId": "35e1a6e790e7a89e56b1fd776773b595",
                "parameters": [
                    {
                        "name": "id_tmdb",
                        "in": "path",
                        "description": "ID do filme no TMDB",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do filme retornados com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "ID do filme",
                                            "type": "integer"
                                        },
                                        "title": {
                                            "description": "Título do filme",
                                            "type": "string"
                                        },
                                        "overview": {
                                            "description": "Resumo do filme",
                                            "type": "string"
                                        },
                                        "poster_path": {
                                            "description": "Caminho do poster do filme",
                                            "type": "string"
                                        },
                                        "release_date": {
                                            "description": "Data de lançamento",
                                            "type": "string",
                                            "format": "date"
                                        },
                                        "vote_average": {
                                            "description": "Nota média do filme",
                                            "type": "number",
                                            "format": "float"
                                        },
                                        "vote_count": {
                                            "description": "Número de votos",
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro ao buscar filme",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "description": "Mensagem de erro",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Favorites",
            "description": "Favorites"
        },
        {
            "name": "Movies",
            "description": "Movies"
        }
    ]
}